+++ test/mocha.opts
@@ -1,4 +1,4 @@
 --require ts-node/register
 --require source-map-support/register
 --recursive
-src/tests.ts src/**/*.test.ts
+test/tests.ts test/**/*.test.ts
